% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAG_Model.R
\name{BAG_Model}
\alias{BAG_Model}
\title{Title}
\usage{
BAG_Model(Data, xvar, yvar)
}
\arguments{
\item{Data}{The name of the Dataset.}

\item{xvar}{X variables.}

\item{yvar}{Y variable.}
}
\value{
The output from  \code{\link{BAG_Model}}.
}
\description{
Title
}
\examples{
\donttest{
yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.BAG <- BAG_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
BchMk.BAG$finalModel
BchMk.BAG$Roc$auc
}
}
