% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined_Performance.R
\name{Combined_Performance}
\alias{Combined_Performance}
\title{Combined Performance of the Data Splits}
\usage{
Combined_Performance(Sub.Est.Mdls)
}
\arguments{
\item{Sub.Est.Mdls}{is the total perfomance of exog.}
}
\value{
The output from  \code{\link{Combined_Performance}}.
}
\description{
Combined Performance of the Data Splits
}
\examples{
\donttest{
sample_data <- sample_data[c(1:750),]
yvar <- c("Loan.Type")
xvar <- c("sex", "married", "age", "havejob", "educ", "political.afl",
"rural", "region", "fin.intermdiaries", "fin.knowldge", "income")
CCP.RF <- Estimate_Models(sample_data, yvar, xvec = xvar, exog = "political.afl",
xadd = c("networth", "networth_homequity", "liquid.assets"),
type = "RF", dnames = c("0","1"))
Sub.CCP.RF <- list (Mdl.1 = CCP.RF$EstMdl$`D.1+networth`,
Mdl.0 = CCP.RF$EstMdl$`D.0+networth`)
CCP.NoCCP.RF <- Combined_Performance (Sub.CCP.RF)
}
}
