% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Models.R
\name{Estimate_Models}
\alias{Estimate_Models}
\title{Title}
\usage{
Estimate_Models(DataSet, yvar, exog = NULL, xvec, xadd, type, dnames)
}
\arguments{
\item{dnames}{}
}
\value{

}
\description{
Title
}
\examples{
CCP.RF.1  <- Estimate_Models(sample_data, yvar=c("Loan.Type"), exog = "political.afl", xvec = m2.xvar0, xadd = "networth", type = "RF", dnames = c("0","1"))

}
