% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_Model.R
\name{GLM_Model}
\alias{GLM_Model}
\title{Title}
\usage{
GLM_Model(Data, xvar, yvar)
}
\arguments{
\item{Data}{The name of the Dataset.}

\item{xvar}{X variables}

\item{yvar}{Y variable}
}
\value{
The output from  \code{\link{GLM_Model}}
}
\description{
Title
}
\examples{
yvar <- c("multi.level")
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
m2.xvar1 <- c(m2.xvar0, "networth")
BchMk.GLM_Model <- GLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
BchMk.GLM_Model$finalModel
BchMk.GLM_Model$Roc$auc

}
