% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\name{hello2}
\alias{hello2}
\title{My Hello2 Word Function}
\usage{
hello2(x)
}
\arguments{
\item{x}{The name of the person to say hi to.}
}
\value{
The output from  \code{\link{print}}
}
\description{
My Hello2 Word Function
}
\examples{
hello2("Seyma")
\dontrun{
hello2("Steve")
}
}
