% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_Model.R
\name{MLM_Model}
\alias{MLM_Model}
\title{Title}
\usage{
MLM_Model(Data, xvar, yvar)
}
\value{
The output from  \code{\link{MLM_Model}}
}
\description{
Title
}
\examples{
yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
BchMk.MLM$finalModel
BchMk.MLM$Roc$auc
}
