% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{My Hello2 Word Function}
\usage{
printer(r, x)
}
\arguments{
\item{r}{what do you want in the second column}

\item{x}{what do you want in the first column}
}
\value{
A tibble
}
\description{
My Hello2 Word Function
}
\examples{
printer(x = rnorm(5), r = rnorm(5))
}
