% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data for analysis.

A dataset containing the prices and other attributes of almost 54,000
diamonds.}
\format{
A \code{data_frame} with 53940 rows and 10 variables:
\describe{
  \item{x1}{price, in US dollars}
  \item{y1}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
sample_data
}
\description{
Sample data for analysis.

A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
