% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAG_Model.R
\name{BAG_Model}
\alias{BAG_Model}
\title{Title}
\usage{
BAG_Model(Data, xvar, yvar)
}
\arguments{
\item{Data}{The name of the Dataset.}

\item{xvar}{X variables}

\item{yvar}{Y variable}
}
\value{
The output from  \code{\link{BAG_Model}}
}
\description{
Title
}
\examples{
yvar <- c("Loan.Type")
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
m2.xvar1 <- c(m2.xvar0, "networth")
BchMk.MLR.1 <- BAG_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
BchMk.MLR.1$finalModel
BchMk.MLR.1$Roc$auc

}
