% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined_Performance.R
\name{Combined_Performance}
\alias{Combined_Performance}
\title{Title}
\usage{
Combined_Performance(Sub.Est.Mdls)
}
\arguments{
\item{Sub.Est.Mdls}{}
}
\value{

}
\description{
Title
}
\examples{
CCP.RF.1  <- ESTIMATE_MODELS(sample_data, yvar=c("Loan.Type"), exog = "political.afl", xvec=m2.xvar0, xadd="networth", type="RF", dnames=c("0","1"))
CCP.GLM.1  <- ESTIMATE_MODELS(CCP, yvar=c("Loan.Type"), exog = "political.afl", m2.xvar0, xadd="networth", type="GLM", dnames=c(0,1))

CCP.MLR.1 <-  RF_Model(CCP, m2.xvar1, yvar)
NoCCP.MLR.1 <-  RF_Model(NoCCP, m2.xvar1, yvar)
Sub.CCP.MLR.RF <- list(Mdl.1 = CCP.MLR.1,  Mdl.2= NoCCP.MLR.1  )
}
