% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined_Performance.R
\name{Combined_Performance}
\alias{Combined_Performance}
\title{Combined Performance of the Data Splits}
\usage{
Combined_Performance(Sub.Est.Mdls)
}
\arguments{
\item{Sub.Est.Mdls}{is the total perfomance of exog.}
}
\value{
The output from  \code{\link{Combined_Performance}}.
}
\description{
Combined Performance of the Data Splits
}
\examples{
\donttest{
library(dplyr)
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex","married","age","havejob","educ","rural","region","income")
m2.xvar1 <- c(m2.xvar0, "networth")   # with "networth"
CCP <- sample_data \%>\% filter(sample_data$political.afl==1)
NoCCP <- sample_data \%>\% filter(sample_data$political.afl==0)
}
}
