% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Models.R
\name{Estimate_Models}
\alias{Estimate_Models}
\title{Results of the Each Data and Data Splits}
\usage{
Estimate_Models(DataSet, yvar, exog = NULL, xvec, xadd, type, dnames)
}
\arguments{
\item{DataSet}{The name of the Dataset.}

\item{yvar}{Y variable.}

\item{exog}{is a vector to be subtract from the calculation.}

\item{xvec}{is a vector of the variables to be used.}

\item{xadd}{is an additional vector to be used.}

\item{type}{can be RF, GLM, MLM, BAG, and GBM.}

\item{dnames}{is the unique values of exog.}
}
\value{
The output from  \code{\link{Estimate_Models}}.
}
\description{
Results of the Each Data and Data Splits
}
\examples{
\donttest{
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex","married","age","havejob","educ","rural","region","income")
CCP.RF <- Estimate_Models(sample_data, yvar = c("Loan.Type"),
exog = "political.afl", xvec = m2.xvar0,
xadd = "networth", type = "RF", dnames = c("0","1"))
}
}
