% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Models.R
\name{Estimate_Models}
\alias{Estimate_Models}
\title{Title}
\usage{
Estimate_Models(DataSet, yvar, exog = NULL, xvec, xadd, type, dnames)
}
\value{
The output from  \code{\link{Estimate_Models}}
}
\description{
Title
}
\examples{
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income") #, "ethnicity" "class.of.HH",
}
