% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_Model.R
\name{GBM_Model}
\alias{GBM_Model}
\title{Title}
\usage{
GBM_Model(Data, xvar, yvar)
}
\arguments{
\item{Data}{The name of the Dataset.}

\item{xvar}{X variables.}

\item{yvar}{Y variable.}
}
\value{
The output from  \code{\link{GBM_Model}}.
}
\description{
Title
}
\examples{
yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:120),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.GBM <- GBM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
BchMk.GBM$finalModel
BchMk.GBM$Roc$auc
}
