% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_Model.R
\name{GLM_Model}
\alias{GLM_Model}
\title{Title}
\usage{
GLM_Model(Data, xvar, yvar)
}
\arguments{
\item{Data}{The name of the Dataset.}

\item{xvar}{X variables.}

\item{yvar}{Y variable.}
}
\value{
The output from  \code{\link{GLM_Model}}.
}
\description{
Title
}
\examples{
yvar <- c("multi.level")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.GLM <- GLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
BchMk.GLM$finalModel
BchMk.GLM$Roc$auc
}
